refactor: move parseObject to common/utils.mjs
